package indextree renamed to radixtree
