+++ internal/radixtree/node.go
@@ -1,10 +1,10 @@
 /*
-Package indextree implements a tree lookup for values associated to
+Package radixtree implements a tree lookup for values associated to
 paths.
 
 This package is a fork of https://github.com/dimfeld/httptreemux.
 */
-package indextree
+package radixtree
 
 import (
 	"errors"
