+++ internal/radixtree/node_benchmark_test.go
@@ -1,4 +1,4 @@
-package indextree
+package radixtree
 
 import (
 	"testing"
